// frontend/src/api/types.ts
export interface Fixture {
    match_id: number;
    kickoff_time: string; // ISO string
    kickoff_timezone: string;
    gameweek: number;
    venue: string;
    home_team_id: number;
    home_team_name: string;
    home_team_abbr: string;
    away_team_id: number;
    away_team_name: string;
    away_team_abbr: string;
}

export interface CompletedFixture extends Fixture {
    home_team_score: number;
    away_team_score: number;
    home_team_redcard: number;
    away_team_redcard: number;
    home_stats: TeamMatchStats;
    away_stats: TeamMatchStats;
    events: MatchEvents;
    home_team_lineup: TeamLineup;
    away_team_lineup: TeamLineup;
    match_report: string;
}

export interface TeamLineup {
    teamId: number;
    lineup: any[];
    players: number[];
    subs: number[];
    formation: string;
}

export interface TeamMatchStats {
    goals: number;
    possessionPercentage: number;
    totalPass: number;
    ontargetScoringAtt: number;
    totalScoringAtt: number;
    expectedGoals: number;
    wonCorners: number;
    fkFoulLost: number; // Fouls conceded
    duelWon: number;
    aerialWon: number;
    // Add more as needed
}

export interface MatchEvents {
    homeTeam: TeamEvents;
    awayTeam: TeamEvents;
}

export interface TeamEvents {
    id: string;
    name: string;
    goals: GoalEvent[];
    cards: CardEvent[];
    subs: SubEvent[];
}